Fix the building of the gem for Ruby versions 2.0.0 and greater
Starting in Ruby 2.0.0, the Gem::Builder module was moved to
Gem::Ext::Builder and the gem build functionality was moved to
Gem::Package. This fix checks the ruby version and if the user is using
Ruby 2.0.0 or greater, we require the Gem::Package module and build the
gem that way.